Implement device switching using the view
NOTE: this causes the actual input device to change,
and therefore all of the readings from the device to
switch to the new device.
HOWEVER, it does not touch the Drive/Arm Controllers,
and as such, loses contact with the original input device.
THIS CODE DOES NOT ALLOW THE USER TO USE DIFFERENT INPUT
DEVICES WITH THE ROVER

TODO: replace the DRIVE/ARM controllers' input device with an
input manager. This will indirectly involve moving the
Controller Modes and all associated functions out of the
individual input devices, and into the input manager