Reworked IDataRouter into INetworkMessenger
The original interface did little besides act as a buffer between the controls and the network. While this would be useful if there were more messages being passed around in our app, in reality we were only ever messaging back and forth between the controls and the network specifically. So it has been reworked as a network messaging package instead, with the dataRouter class itself now merged into NetworkManager