Refactored rove protocol objects into centralized rovecomm object
It's much more centralized than how it was previously, now all the logic for rovecomm related activities are kept in one location. As well, it keeps it more in line with how the rest of our systems have become, greatly adding to RED's clarity for onboarding.

Network layer could still use a bit of work itself; acks have been deprecated for now due to being buggy