Attempt 1
Attempted to create an instance of input manager, and then pass "this" onto its children (main window down to arm, science, and autonomy).  Then those views set their instance of input manager equal to the parent's input manager. This created an error in the constructor for the the input manager.