Update for 2020 Rovecomm IDs
I think everything is updated properly according to the manifest in github. Notes: Nav has nothing. Drive and Lighting are seperate but the same board ids. Some of the descriptions suck either on my end or on the document end, but I believe these to be all correct. No errors are implemented.