Fixed 3D model rotation
We had the rover set to rotate to the pitch, yaw, and roll of the rover. Moving to the yaw made it impossible to figure out what was being depicted (as noted, just implemented incorrectly). I also realized it makes more sense to just send rotation packets when we get the packets rather than when the variables are updated. This calls the function less. Finally, the orientation of the N3 on the rover means that -pitch and -roll is what the rover should be rotated based off of.