Converting chemical control to int16s
According to the board software, it claims that the chemical motors are looking for an int16. I assume it wants a number between 0 and 1000, like just the positive half of a drive motor. The rovecomm manifest originally said it should just be looking for a 0 or 1 byte, and while this has been updated, exactly what it is looking for hasn't changed.