Begin implementing tasks for the timer
There was no notion of tasks prior to this commit. Instead, a single
allotted time was passed into the Timer component as a prop. For one,
the time in this case cannot be modified, as props are read-only in
React, but additional times cannot be added either.

Now, tasks are managed in the state of the Timer component, hardcoded
for the time being. In the future, this allows for several benefits,
including:
    - Ability to package times with respective task names
    - Option to add, remove, and/or modify a task, name or time