additional outlining
Styling for the feature is further roughed out with most of the variables being declared in the IState interface. The styling is definately bloated, but I'd like advising on where to go from here, as there are properties of the C# software that I'm unsure of in replicating the layout of the feature.