Implemented basic custom packet feature
This is a really trimmed down version of rovecomm tester. It has drop downs to select the board you want to send a command to, the command you want to send, and then enter the data in and send it. Doesn't support custom ips, dataIds, anything not matching the manifest, controllers, etc.