Began total rework (won't take long)
In the midst of learning how to integrate rovecomm data, it was determined 500 lines of code was too long for this feature. This commit contains the beginning of an array based data structure that should correctly integrate rovecomm data via event listeners. Current direction is pretty solid and should be fairly quick to implement.