Add React Table Styles
Needs to be added to the app.html and linked the the ReactDOM.createportal(). There is a current problem in that the linking to the new window loses all reference to files, so we have to link to a website, which is not a long term solution. Also fixed the babel css connection