Make 3D rover more dynamically sized
Now, when a new window is created we add it to an object to be exported. This object can then be referred to later by other classes to "look up" the size of their window or div, or anything else you may want to do with other screens document or window variables. This may have important implications for other components and issues. Note: on the first pass, window hasn't actually been updated with the reference to the monitor yet, as the monitor has to be rendered before the reference can be made. So this solution is still a little hacky and will likely require a timeout to resize 100ms after componentDidMount (or something similar)