Fix screenshotting issues across the board
Part of the reason cameras and graphs could not be screenshot is because we were capturing the wrong pixels (and still may be for cameras). Now we are not using the default window but the actual window we find the element in and then properly offsetting by the top border and negative margin.