Fix sensor graph crosshair
OnNearestX (which populates crosshair values) was finding the index of the nearest value in methane and then grabbing that value from every sensor. However, methane updates every 21 seconds, oxygen every 2 seconds, and co2 every second, making our lists all kinds of different sizes. To make this cleaner, we changed crosshair values to a dictionary with the sensor as the key value pair and we call the crosshair for each line series present.