Properly normalize all data graphs
The all graphs screen was empty because the first entry of each series was NaN, as we are normalizing as a function of the greatest previous input, which for the first element is always 0 and results in a div by zero error. So now if the max is 0 we do some special checking.
(Also remove some annoying console logs from arm that made debugging this difficult)