Fixed Issue Where Attempting to Delete a Task Caused Crash
Currently, the JSON gets saved every time a task is deleted, but for a currently unknown reason, when a mission is deleted, the JSON is only saved if a second mission is subsequently deleted. removeInstance() is supposed to save the JSON every time the function is called, but it appears to save the JSON on the second call if what's getting deleted are missions.