Fleshed out submenu base for the modal containing controller layouts
At the moment, there's a missing image and some arbitrary text to mark a preview slot. Thomas advised making the layout a list at first before I make cool little graphics with the xbox controllers and flight stick. Whether the layout is a list or a graphic, though, both will require importing the controller JSON or the controllerinput JSON, which will be worked on next.