Minor UI Changes #3
I am at the point where the mission title and target mission time are aligned correctly on the menu. In addition to that, the subscreen inside the menu that contains the task list and their respective target times. I am having issues aligning the difference time evenly among the task list, but that should be solved soon.