Nearly Finished UI/Time Difference Functionality
The UI for the time split is nearly finished, all that is left component-wise is the button to store the differences in a JSON file. Other than that, the UI reflects what was designed in Figma, and any additional changes needed would most likely be only cosmetic. I managed to be able to print the "delta" variable in accordance to each task on the time split, but there are a few hiccups that can occur when messing with the UI which can be ironed out later.