Small Time Split UI Updates/Hiccup Fixes
I added the large button that saves the time differences to a JSON file added at the bottom of the time split menu. For the time being, the differences will be saved to the TaskList.json file. I attempted to fix some time difference functionality issues, such as resetting a time difference when going to a previous task or just resetting all the time differences in general when calling reset(). I also attempted to have the timer record the difference for the last task of the last mission by allowing the user to press "Next Task" one last time on the last task. I had the UI change "Next Task" to "End Mission" when reaching the final task.