Post Code Review Changes
Code was cleaned up a little bit, a bit of redundant code was removed in isTaskListEmpty(). The delta that would be printed out over the task progress bar was deleted, since we determined that it would be more beneficial to move delta over to the time split modal. Overall, very minor changes.