(Nearly) Finishing Touches to Time Split
In terms of functionality, time split runs exactly how it should. The menu shows a concise overlay of the current mission's tasks and task completion differences. The split menu also saves the difference data to a .txt file with the most recent save appearing at the top. The only changes I see being made in the future are small cosmetic changes to the modal and the code itself and a potential fix to the task/mission deletion bug.