Fixes and small changes
Fixed the rock database so no more duplicates are happening, and the uv/white lights toggles were moved to Heater.tsx (aka science hardware), and a fix was attempted on Timer. I'm currently unable to replicate the bug that happened during the task sim a few days ago, but the error seemed to be sourced from non-unique html element IDs. I attempted to address this by assigning unique IDs programmatically.